requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
bitarray==2.9.2
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==2.0.0
cytoolz==0.12.3
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
eth-account==0.13.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
frozenlist==1.4.1
hexbytes==1.2.1
idna==3.7
inflection==0.5.1
multidict==6.0.5
packaging==24.1
parsimonious==0.10.0
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
setuptools==73.0.1
sqlparse==0.5.1
toolz==0.12.1
types-requests==2.32.0.20240712
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
web3==7.0.0
websockets==13.0
wheel==0.37.1
yarl==1.9.4
